+++ src/main/java/com/flightstats/hub/model/Traces.java
@@ -1,32 +1,102 @@
 package com.flightstats.hub.model;
 
+import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
+import org.joda.time.DateTime;
 import org.slf4j.Logger;
 
-import java.util.SortedSet;
+import java.util.*;
 
-public interface Traces {
+public class Traces {
 
-    void end();
+    private long start = System.currentTimeMillis();
+    private long end;
+    private final String id = UUID.randomUUID().toString();
+    private final List<Trace> traces = Collections.synchronizedList(new ArrayList<>());
 
-    long getTime();
+    public Traces() {
+    }
 
-    void add(Trace trace);
+    public Traces(Object... objects) {
+        add(objects);
+    }
 
-    void add(Object... objects);
+    public void end() {
+        end = System.currentTimeMillis();
+        add("cend");
+    }
 
-    void add(String string, SortedSet sortedSet);
+    public long getTime() {
+        return end - start;
+    }
 
-    void setStart(long start);
+    public void add(Trace trace) {
+        traces.add(trace);
+    }
 
-    long getStart();
+    public void add(Object... objects) {
+        traces.add(new Trace(objects));
+    }
 
-    String getId();
+    public void add(String string, SortedSet sortedSet) {
+        if (sortedSet.isEmpty()) {
+            add(string, "empty set");
+        } else {
+            add(string, sortedSet.size(), sortedSet.first(), sortedSet.last());
+        }
+    }
 
-    void logSlow(long millis, Logger logger);
+    public void setStart(long start) {
+        this.start = start;
+    }
 
-    void output(ObjectNode root);
+    public long getStart() {
+        return start;
+    }
 
-    void log(Logger logger);
+    public String getId() {
+        return id;
+    }
+
+    public void logSlow(long millis, Logger logger) {
+        long processingTime = System.currentTimeMillis() - start;
+        if (processingTime >= millis) {
+            String output = getOutput(logger);
+            logger.info("slow processing of {} millis. trace: {}", processingTime, output);
+        }
+    }
+
+    public void log(Logger logger) {
+        String output = getOutput(logger);
+        logger.info("trace: {}", output);
+    }
+
+    private String getOutput(Logger logger) {
+        try {
+            traces.add(new Trace("logging"));
+            String output = "\n\t";
+            synchronized (traces) {
+                for (Trace trace : traces) {
+                    output += trace.toString() + "\n\t";
+                }
+            }
+            return output;
+        } catch (Exception e) {
+            logger.warn("unable to log {} traces {}", traces);
+            return "unable to output";
+        }
+    }
+
+    public void output(ObjectNode root) {
+        root.put("id", id);
+        root.put("start", new DateTime(start).toString());
+        ArrayNode traceRoot = root.putArray("trace");
+        synchronized (traces) {
+            for (Trace trace : traces) {
+                traceRoot.add(trace.toString());
+            }
+        }
+    }
+
 
 }
